+++ packages/util-crypto/src/nacl/sign.spec.ts
@@ -3,13 +3,13 @@
 // of the Apache-2.0 license. See the LICENSE file for details.
 
 import { stringToU8a } from '@polkadot/util';
-import ed25519 from '@polkadot/wasm-dalek-ed25519';
+import { waitReady } from '@polkadot/wasm-crypto';
 
 import { naclSign, naclKeypairFromSeed } from '.';
 
 describe('naclSign', () => {
   beforeEach(async () => {
-    await ed25519.waitReady();
+    await waitReady();
   });
 
   it('returns a valid signature for the message', () => {