+++ packages/util-crypto/src/nacl/verify.ts
@@ -3,7 +3,7 @@
 // of the Apache-2.0 license. See the LICENSE file for details.
 
 import nacl from 'tweetnacl';
-import { isReady, verify } from '@polkadot/wasm-dalek-ed25519';
+import { isReady, ed25519Verify } from '@polkadot/wasm-crypto';
 
 /**
  * @name naclSign
@@ -21,6 +21,6 @@ import { isReady, verify } from '@polkadot/wasm-dalek-ed25519';
  */
 export default function naclVerify (message: Uint8Array, signature: Uint8Array, publicKey: Uint8Array): boolean {
   return isReady()
-    ? verify(signature, message, publicKey)
+    ? ed25519Verify(signature, message, publicKey)
     : nacl.sign.detached.verify(message, signature, publicKey);
 }